// src/types/game.ts
export interface Game {
    gameID: string;
    league: string;
    week: number;
    year: number;
    awayTeamID: string;
    homeTeamID: string;
    date: Date;
    time: string;
    tvCoverage: string;
    stadium: string;
    city: string;
    state: string;
    gameFinished: boolean;
    cbsCode: string;
    cbsAwayMoneyline: string;
    cbsHomeMoneyline: string;
    cbsAwaySpread: string;
    cbsHomeSpread: string;
    cbsOverUnder: string;
    cbsAwayWinPercentage: string;
    cbsHomeWinPercentage: string;
    espnCode: string;
    espnAwayMoneyline: string;
    espnHomeMoneyline: string;
    espnAwaySpread: string;
    espnHomeSpread: string;
    espnOverUnder: string;
    espnAwayWinPercentage: string;
    espnHomeWinPercentage: string;
    foxCode: string;
    foxAwayMoneyline: string;
    foxHomeMoneyline: string;
    foxAwaySpread: string;
    foxHomeSpread: string;
    foxOverUnder: string;
    foxAwayWinPercentage: string;
    foxHomeWinPercentage: string;
    vegasCode: string
    vegasAwayMoneyline: string;
    vegasHomeMoneyline: string;
    vegasAwaySpread: string;
    vegasHomeSpread: string;
    vegasOverUnder: string;
    vegasAwayWinPercentage: string;
    vegasHomeWinPercentage: string;
    awayQ1BoxScore: number;
    awayQ2BoxScore: number;
    awayQ3BoxScore: number;
    awayQ4BoxScore: number;
    awayOvertimeBoxScore: number;
    awayTotalBoxScore: number;
    homeQ1BoxScore: number;
    homeQ2BoxScore: number;
    homeQ3BoxScore: number;
    homeQ4BoxScore: number;
    homeOvertimeBoxScore: number;
    homeTotalBoxScore: number;
}